Allow arbitrary handlers to support more runtimes. (#22)
GCP Cloud functions have an HTTP deployment that requires a different function signature, seen here: https://cloud.google.com/functions/docs/writing#functions-writing-helloworld-http-python

This PR allows for lambdex to support arbitrary handlers, which in turn allows for signatures such as GCP's. In order to allow GCP Cloud Function users to test their functions using lambdex, we needed to add the ability for the test function to call this specific type.

If used, the user will need to have flask installed, since the test will create a flask request that is sent to the function.

I also took the opportunity to create another example that is specifically for a GCP HTTP Cloud Functions.