Fix main CI (#25)
Several deps had drifted into territory that led to errors with Python
2.7 and 3.6 shards; those are now constrained appropriately.

In addition, Pex 2.1.110 brought further hardening to the PEX boot 
`sys.path` scrubbing process which revealed a latent bug in the way the
Lambdex PEX used in tests was being constructed. That is now fixed with
a dedicated script to create that Lambdex PEX and ensure it's built with
the intended version of Pex.